requirements.txt
click==8.1.3
Flask==2.2.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
numpy==1.19.2
pandas==1.1.5
pillow==6.2.1
flasgger==0.9.5
torch==1.10.2
torchvision==0.11.3
segmentation_models_pytorch==0.3.1